all: remove duplicate and simplify code
